% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kosac.R
\name{get_kosac}
\alias{get_kosac}
\title{get_kosac}
\usage{
get_kosac(lexicon = c("polarity", "intensity", "subj_polarity",
  "subj_type"))
}
\arguments{
\item{lexicon}{"polarity" is default.}
}
\description{
get_kosac
}
