% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSACR-package.R
\docType{package}
\name{KOSACR-package}
\alias{KOSACR}
\alias{KOSACR-package}
\title{KOSACR: Korean Sentiment Dictionary from KOSAC}
\description{
Korean Sentiment Dictionary Packages
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/KOSACR}
  \item Report bugs at \url{https://github.com/mrchypark/KOSACR/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
