% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiments.R
\docType{data}
\name{sentiments}
\alias{sentiments}
\title{sentiments}
\format{tibble with colum}
\source{
http://word.snu.ac.kr/kosac/
}
\usage{
sentiments
}
\description{
sentiments
}
\examples{
  "sentiments"
}
\keyword{datasets}
